refactor(Calculator): refactoring
use it syntax instead of lambda syntax
use noun instead of verb
Regex object to const value
change magic literal to const
