+++ src/main/kotlin/calculator/PositiveNumbers.kt
@@ -1,21 +1,17 @@
 package calculator
 
-data class PositiveNumbers(private val value: List<Int>) {
+data class PositiveNumbers(val value: List<Int>) {
+    init {
+        if (hasNegativeNumber(value)) {
+            throw RuntimeException("음수는 입력할 수 없습니다.")
+        }
+    }
+
     fun sum(): Int {
         return value.sum()
     }
 
-    companion object {
-        fun parse(numbers: List<Int>): PositiveNumbers {
-            if(hasNegativeNumber(numbers)) {
-               throw RuntimeException("음수는 입력할 수 없습니다.")
-            }
-
-            return PositiveNumbers(numbers)
-        }
-
-        private fun hasNegativeNumber(numbers: List<Int>): Boolean {
-            return numbers.any { number -> number < 0 }
-        }
+    private fun hasNegativeNumber(numbers: List<Int>): Boolean {
+        return numbers.any { number -> number < 0 }
     }
 }
